+++ NEWS.md
@@ -2,7 +2,7 @@
 **NSGEV** Package News
 ======================
 
-# New in version 0.2.0 (branch `influence`
+# New in version 0.2.0 (branch `influence`)
 
 ## Enhancements
 
@@ -14,8 +14,22 @@
   
 ## Bug fixes
  
-- The generalised residuals computed with `"gumbel"` were 
-  actually *reversed* gumbel. 
+- The generalised residuals computed with `"gumbel"` were actually
+  following the *reversed* Gumbel, not the Gumbel. New (experimental)
+  choice `"gev4"` for the target distribution.
+  
+- When `TVGEV` was used with `optim = "nloptr"`, the choice `parTrack
+  = TRUE` did not work as expected.
+
+## Changes
+ 
+- The (non-exported) `MLE.TVGEV` function now has a `trace` argument
+  which is set to the value of `trace` argument of `TVGEV`.  When
+  `estim = "nloptr"` and a value `> 2` is given the values of the 
+  parameter and of the objective are traced.
+  
+- The Hessian of the negalive log-likelihood is now exact and no longer 
+  computed with `optimHessian`.
 
 # New in version 0.2.0